Fix from_canonical_usize

+++ field/src/field.rs
@@ -144,7 +144,7 @@ pub trait Field:
 pub trait PrimeField: Field + Ord {
     fn from_canonical_u32(n: u32) -> Self;
     fn from_canonical_u64(n: u64) -> Self;
-    fn from_canonical_usize(n: u64) -> Self;
+    fn from_canonical_usize(n: usize) -> Self;
 
     fn from_wrapped_u32(n: u32) -> Self;
     fn from_wrapped_u64(n: u64) -> Self;
